requirements.txt
django-cors-headers==2.5.2
django-filter==1.1.0
djangorestframework==3.9.2
Pillow==5.4.1
numpy==1.16.2
pandas==0.24.2
tqdm==4.31.1
scikit-image==0.15.0
filetype==1.0.5
pydub==0.23.1
python-docx==0.8.10
pdfminer3k==1.3.1
pyocr==0.7.2
django-sortedm2m==2.0.0
moviepy==1.0.0
redis==3.3.11
django-redis==4.10.0
rq==1.1.0
django-rq==2.1.0
django-background-tasks==1.2.0